Step2:리팩터링 반영
